PUPPET-181 Better separation of package and archive resource properties
By re-using the package resource properties for the archive one as well, it is quite hard to support OS-specific default values in one case and provide a different default value in the other case whilst allowing to user to override the default as well. Splitting them makes it much clearer which one is used for what.