Add profile parameter to specifiy include/exclude in verbatim
If there are already sources for rsync-style include/exclude
specifications, it is hard to adapt them to the plain array
style of the `include_filelist` and `exclude_filelist` parameters.

This commit adds the ability to set the content verbatim, while
forcing the exclusive use of this feature. It is meant as a power
user feature for cases, where the include/exclude specification
is of a more complex nature.